firebase.config.jsファイルを作成し、WebAPIキーを別ファイルで管理。axiosパッケージをアンインストール
