+++ package.json
@@ -12,16 +12,13 @@
     "firebase": "^8.2.4",
     "vue": "^3.0.0",
     "vue-auth": "^1.0.0-alpha.1",
-    "vue-axios": "^3.2.2",
-    "vue-router": "^4.0.3",
-    "vue3-use-axios": "^0.2.2"
+    "vue-router": "^4.0.3"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
     "@vue/compiler-sfc": "^3.0.0",
-    "axios": "^0.21.1",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^7.0.0-0"
